+++ src/mlde_data/variable/__init__.py
@@ -1,0 +1,18 @@
+from string import Template
+from pathlib import Path
+import yaml
+
+
+def load_config(config_path: Path, scale_factor: str, domain: str, size: int):
+    """Load configuration for a creating a variable from a YAML file."""
+
+    with open(config_path, "r") as config_template:
+        d = {
+            "scale_factor": scale_factor,
+            "domain": domain,
+            "size": size,
+        }
+        src = Template(config_template.read())
+        result = src.substitute(d)
+        config = yaml.safe_load(result)
+    return config
